fix #15: Add toggl parser script
Have now created a script that can be used to parse .csv files from
toggle, converting the time to minutes, and printing the total time used
for each issue in a csv format.

Also added a file with time entries for testing purposes.

ref #11: Should not need this anymore, as we can use the data from
toggle, instead of data from commit messages.