ref #28: Remove non async step from lifecycle
Removed the paragraph stating that you could call initialize on an
object right after you have gotten the handle from the EntityManager.
this feature is decided to be removed, as it would be inconsistent to
have one non async part of the lifecycle.
Also it is harder to do non async.

Also did some minor formatting.